Improve utilities export 2
- dbadmin/administrar_ex.php renamed to inicio_bd.php to avoid confusion. Only function inicializar was used.
- inicio_bd.php with only 1 function and some improvements. Works but not decent
- menu_bar.php: invokes inicio_bd.php in stead of administrar_ex.php
- vmx_fullinv.php: Show correct heading and backbutton for second invocation
